singly list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 size_t print_list(const list_t *h)
 {
 	unsigned int i;
-	unsigned int result;
+	size_t result;
 
 	for (i = 0; h; i++)
 	{
